Added composer scripts for PHPCS.
Windows, Mac, and Linux friendly script syntax to run PHPCS from composer's
vendor/bin.

Locally, we can now run:

`composer phpcs-dev`

which runs the PHPCS sniffer on the `dev` folder files only.

or...

`composer run-phpcs`

which runs all of the PHPCS scripts.

Once we add tests, then we'll have the ability to run PHPCS on the theme's
code, the test suites, or both.