Optimize gathering and enqueuing for custom fonts.
to this approach are:

1. The intent is clearly communicated.
2. The empty string is not processed through esc_url_raw().  That gives
us a little decrease in memory and processing time.
3. We don't have to pre-declare $fonts_url.  That gives us a little
decrease in memory and processing time too.

`wprig_fonts_url()` returns an empty string when no fonts are turned
on.  When that happens, we do not want to enqueue, as there's nothing
to be enqueued.

This change checks if we got something back.  If yes, then it enqueues.

The results are:

1. a small performance increase when no fonts are enabled.
2. wprig is more robust.

`in_array()` is slower than `empty()`.  Here we can simply check if
anything got added to the array.  If yes, then we know that one of the
font families was turned on.  If no, then bail out.

Result: A little bit faster.