Exit the prodFinish task early if compress is false
The task `prodFinish` creates the production `.zip` archive. Creation of the production `.zip` archive is configurable via `config.export.compress`, which defaults to `true`.

There is a flaw in the logic of the `prodFinish` task. All the production files are brought in with `src`, conditionally turned into a `.zip` based on `config.export.compress`, then written to the directory above the development/production themes, e.g. `wp-content/themes`. This causes all the production files to be written to `wp-content/themes` when `config.export.compress` is `false`.

This change exits the `.zip` creation process early, rather than relying of a conditional in the middle of the task, if `config.export.compress` is `false` to prevent this unwanted behaviour.