Add a source string replacement gulp task
The task `sourceStringReplacement`, which can be run with `npm run dev -- sourceStringReplacement`, will run string replacement in the source directory, as opposed to the production directory.

This is not encouraged but can be done to essentially eject from WP Rig as use the source theme as-is instead of needing to build a production theme.