Combine style tasks into a single stream
Use [`pipeline` from `mississippi`](https://www.npmjs.com/package/mississippi#pipeline) to convert the tasks in before replacement and after replacement into a single stream.

This allows a much cleaner use of `pump`.

The tasks to run before and after replacement are now in exported functions so that they can be tested.