Single stream gulp tasks (#397)
Use [`pipeline` from `mississippi`](https://www.npmjs.com/package/mississippi#pipeline) to convert the tasks in before replacement and after replacement into a single stream. This allows a much cleaner use of `pump`.

The tasks to run before and after replacement are now in exported functions so that they can be re-used. This is a pre-requisite for writing tests for gulp tasks (#371).

Props @phated!