Always copy, and run string replacement, on style.css
Adds a new task `prodStringReplace`, which copies files requirted for the bundled theme but not created in a build step to the production theme and runs string replacement on them.

Rename `prodFinish` to `prodCompress` as the task creates the `.zip` of the production bundle.

Remove `style.css` from `filesToCopy` in the default config as it is now handled with the new `prodStringReplace` task.