Initialize navigation even if DOMContentLoaded has fired
Because the `navigation.min.js` script is loaded asynchronously in
the browser, it's possible that the `DOMContentLoaded` event has
already fired by the time the browser processes the script.

We can check the state of the DOM when the script loads to determine
if the event should be used or if the script should initialize
immediately.