Clarify naming of wprig development folder in the installation process (#441)
* Update README.md

* Update README.md

* Update README.md

* Update README.md

Removed the link to the production process further on as I do not actually think this is necessary.

* Update README.md

* Update README.md

I liked Benoit's naming suggestion.  I also tried to simplify the language as much as possible.   I also took out the reference to npm run bundle, confusing if someone is brand new to working with the theme in my view.