Use nodir:true in gulp src when creating the .zip
See #594. gulp `src` passes directories by default. However, on Windows and Linux the permissions are different so if the prod bundle is created on a Windows machine and deployed to a Linux server there are issues. Setting `nodir: true` in gulp `src` is a workaround suggested [here](https://github.com/sindresorhus/gulp-zip/issues/64#issuecomment-191659258).