After discussion, decided to leave the is_home() call in front-page.php, but change the css name to home.css.
Added some comments on what is_home is checking and why the particular css
files are chosen.

front-page.php is used as the front page of the website, regardless of whether
the front page is a static page or the blog index (is_home()). In the case
where it's the blog index (i.e., the blog home page) we want to use the grid
layout.

Ideally, the name of the php file and the name of the css file would match,
so one could argue we should have a home.php in the optional folder, and
remove the front-page.php example*. However, front-page.php provides a good
example of a conditional check to include different css files based on some
dependency.

* There's agreement that we don't want a static front-page custom example.